+++ src/lib/libast/sfio/sfhdr.h
@@ -334,7 +334,7 @@ struct _fmt_s
 
 	char*		oform;		/* original format string	*/
 	va_list		oargs;		/* original arg list		*/
-	int		argn;		/* number of args already used	*/
+	ssize_t		argn;		/* number of args already used	*/
 	Fmtpos_t*	fp;		/* position list		*/
 
 	Sffmt_t*	ft;		/* formatting environment	*/
@@ -346,7 +346,7 @@ struct _fmtpos_s
 {	Sffmt_t	ft;			/* environment			*/
 	Argv_t	argv;			/* argument value		*/
 	int	fmt;			/* original format		*/
-	int	need[FP_INDEX];		/* positions depending on	*/
+	ssize_t	need[FP_INDEX];		/* positions depending on	*/
 };
 
 #define LEFTP		'('
@@ -632,7 +632,7 @@ typedef struct _sftab_
 	int		(*sf_cvinitf)(void);	/* initialization function	*/
 	int		sf_cvinit;		/* initialization state		*/
 	Fmtpos_t*	(*sf_fmtposf)(Sfio_t*,const char*,va_list,Sffmt_t*,int);
-	char*		(*sf_fmtintf)(const char*,int*);
+	char*		(*sf_fmtintf)(const char*,ssize_t*);
 	float*		sf_flt_pow10;		/* float powers of 10		*/
 	double*		sf_dbl_pow10;		/* double powers of 10		*/
 	Sfdouble_t*	sf_ldbl_pow10;		/* Sfdouble_t powers of 10	*/
@@ -742,7 +742,7 @@ extern int		_sfpclose(Sfio_t*);
 extern int		_sfexcept(Sfio_t*, int, ssize_t, Sfdisc_t*);
 extern Sfrsrv_t*	_sfrsrv(Sfio_t*, ssize_t);
 extern int		_sfsetpool(Sfio_t*);
-extern char*		_sfcvt(void*,char*,size_t,int,int*,int*,int*,int);
+extern char*		_sfcvt(void*,char*,size_t,ssize_t,int*,int*,ssize_t*,ssize_t);
 extern char**		_sfgetpath(char*);
 
 extern Sfextern_t	_Sfextern;
